fix bug #16
